requirements.txt
openai==1.12.0
rich==13.7.0
tiktoken==0.6.0
pygame==2.5.2
discord.py==1.7.3
discord.py[voice]==1.7.3
PyNaCl==1.4.0
python-dotenv==1.0.1
